% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.simMat.R
\name{as.SimilarityMatrix}
\alias{as.SimilarityMatrix}
\title{Convert a matrix to a similarity matrix.
\code{as.SimilarityMatrix} convert an adjacency matrix to
a similarity matrix.}
\usage{
as.SimilarityMatrix(mat)
}
\arguments{
\item{mat}{a symmetric adjacency matrix}
}
\value{
a similarity matrix.
}
\description{
Convert a matrix to a similarity matrix.
\code{as.SimilarityMatrix} convert an adjacency matrix to
a similarity matrix.
}
\details{
It is required to use \code{as.simMat} to transform your
raw edgelist or raw matrix into a matrix of \code{similarityMatrix}
object before using other functions to find clusters.
}

