% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.simMat.R
\name{as.simMat}
\alias{as.simMat}
\title{convert to a matrix of \code{similarityMatrix} class}
\usage{
as.simMat(Data, weighted = FALSE)
}
\arguments{
\item{Data}{either a dataframe or a matrix, representing raw interactions using either an edgelist or a matrix.
Frequency of interactions for each dyad can be represented either by multiple occurrences of the dyad for a 2-column edgelist, or
by a third column specifying the frequency of the interaction for a 3-column edgelist.}

\item{weighted}{If the edgelist is a 3-column edgelist in which weight was specified by frequency, use \code{weighted = TRUE}.}
}
\value{
a named matrix with the \code{[i,j]}th entry equal to the number of times \code{i} grooms \code{j}.
}
\description{
\code{as.simMat} convert an edgelist or a raw matrix to a similarity matrix whose values range from 0 to 1 of \code{similarityMatrix} class.
}
\details{
It is required to use \code{as.simMat} to transform your
raw edgelist or raw matrix into a matrix of \code{similarityMatrix} object before
using other functions to find clusters.

Note, when using a 3-column edgelist (e.g. a weighted edgelist) to represent raw interactions, each dyad must be unique. If more than one rows are found with the same Initiator and recipient,
sum of the frequencies will be taken to represent the freqency of interactions between this unique dyad. A warning message will prompt your attention to the accuracy of your raw data when duplicated dyads were found in a three-column edgelist.
}
\examples{
simmatrix <- as.simMat(myData)
}

