% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getEigenvalues.R
\name{getEigenvalueList}
\alias{getEigenvalueList}
\title{generate eigenvalues for all ensemble matrices
\code{getEigenvalueList} get eigenvalues from ensemble matrices}
\usage{
getEigenvalueList(EnsList)
}
\arguments{
\item{EnsList}{a list of ensemble matrices}
}
\value{
a list of \code{eigenvalues} for each of the ensemble matrix in the ensemble matrices list.
#
}
\description{
generate eigenvalues for all ensemble matrices
\code{getEigenvalueList} get eigenvalues from ensemble matrices
}
\examples{
#Sim <- as.simMat(myData)
#temperatures <- temperatureSample(start = 0.01, end = 20, n = 20, method = 'random')
#\dontrun{
## for illustration only. skip CRAN check because it ran forever.
#Ens_list <- getEnsList(Sim, temperatures, MaxIt = 1000, m = 5)
#}
#\dontshow{
## for CRAN check only
#Ens_list <- getEnsList(Sim, temperatures, MaxIt = 5, m = 5)
#}
#eigenvalue_list <- getEigenvalueList(Ens_list)
}

