% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getEns.R
\name{getEns}
\alias{getEns}
\title{generate ensemble matrix
\code{getEns} get ensemble matrix from given similarity matrix and temperature}
\usage{
getEns(simMat, temperature, MaxIt = 1000, m = 5)
}
\arguments{
\item{simMat}{a similarity matrix}

\item{temperature}{a numeric vector of length 1, indicating the temperature used to transform the similarity matrix to ensemble matrix}

\item{MaxIt}{parameter used in EstClust function [Read EstClust function to further understand MaxIt].}

\item{m}{parameter used in EstClust function [Read EstClust function to further understand MaxIt].}
}
\value{
a matrix.
}
\description{
generate ensemble matrix
\code{getEns} get ensemble matrix from given similarity matrix and temperature
}
\details{
This function involves two steps.
It first generate similarity matrices of different variances
by taking the raw similarity matrix to the power of each
temperature. Then it called the function \code{EstClust} to perform random walks in the network to identify clusters.
}

