% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSim.R
\name{GetSim}
\alias{GetSim}
\title{\code{GetSim} get similarity matrix from a distance matrix}
\usage{
GetSim(D, T)
}
\arguments{
\item{D}{A distance matrix}

\item{T}{Temperature. \code{\link{temperatureSample}}}
}
\description{
\code{GetSim} get similarity matrix from a distance matrix
}
\details{
the similarity matrix is calculated at each temperature T.
}

