% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monkeyGrooming.R
\docType{data}
\name{monkeyGrooming}
\alias{monkeyGrooming}
\title{Grooming network data}
\format{A data frame with 1595 rows and 2 variables:
\describe{
  \item{Initiator}{Grooming Initiator ID}
  \item{Recipient}{Grooming Recipient ID}
  \item{Groom}{Grooming Frequency}
  ...
}}
\usage{
monkeyGrooming
}
\description{
A dataset containing grooming edgelist among monkeys.
}
\keyword{datasets}

