% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrees.R
\name{plotCLUSTERS}
\alias{plotCLUSTERS}
\title{generate tree plots for each ensemble matrix
\code{plotCLUSTERS} plot all cluster trees}
\usage{
plotCLUSTERS(EnsList, mfrow, mar = c(1, 1, 1, 1), line = -1.5, cex = 0.5,
  ...)
}
\arguments{
\item{EnsList}{a list in which elements are ensemble matrices.}

\item{mfrow}{A vector of the form \code{c(nr, nc)} passed to \code{\link{par}}.}

\item{mar}{plotting parameters with useful defaults (\code{\link{par}})}

\item{line}{plotting parameters with useful defaults (\code{\link{par}})}

\item{cex}{plotting parameters with useful defaults (\code{\link{par}})}

\item{...}{further plotting parameters}
}
\value{
a \code{pdf} file in the working directory containing all tree plots.
}
\description{
generate tree plots for each ensemble matrix
\code{plotCLUSTERS} plot all cluster trees
}
\details{
When deciding parameters for mfrow,
one should take into considerations size of the plotting device and number of cluster plots.
For example, there are 20 cluster plots, mfrow can be set to \code{c(4, 5)} or \code{c(2, 10)}
depending on the size and shape of the plotting area.
}
\examples{
Sim <- as.simMat(myData)
temperatures <- temperatureSample(start = 0.01, end = 20, n = 20, method = 'random')
\dontrun{
# for illustration only. skip CRAN check because it ran forever.
Ens_list <- getEnsList(Sim, temperatures, MaxIt = 1000, m = 5)
}
\dontshow{
# for CRAN check only
Ens_list <- getEnsList(Sim, temperatures, MaxIt = 5, m = 5)
}
plotCLUSTERS(EnsList = Ens_list, mfrow = c(2, 10), mar = c(1, 1, 1, 1))

}

