% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotEigenvalues.R
\name{plotMultiEigenvalues}
\alias{plotMultiEigenvalues}
\title{plot eigenvalues
\code{plotMultiEigenvalues} plot eigen values into a ".pdf" file.}
\usage{
plotMultiEigenvalues(eigenvalue_list, name = "eigenvaluePlots")
}
\arguments{
\item{eigenvalue_list}{a list in which elements are numeric vectors representing eigenvalues.}

\item{name}{a character vector of length 1. name of the exported pdf}
}
\value{
a pdf file in the working directory containing all eigenvalue plots
}
\description{
plot eigenvalues
\code{plotMultiEigenvalues} plot eigen values into a ".pdf" file.
}
\examples{
Sim <- as.simMat(myData)
temperatures <- temperatureSample(start = 0.01, end = 20, n = 20, method = 'random')
\dontrun{
# for illustration only. skip CRAN check because it ran forever.
Ens_list <- getEnsList(Sim, temperatures, MaxIt = 1000, m = 5)
}
\dontshow{
# for CRAN check only
Ens_list <- getEnsList(Sim, temperatures, MaxIt = 5, m = 5)
}
eigenvalue_list <- getEigenvalueList(Ens_list)
plotMultiEigenvalues(eigenvalue_list, "eigenvaluePlots")
}

