% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotTrees.R
\name{plotTheTree}
\alias{plotTheTree}
\title{generate tree plots for selected ensemble matrix
\code{plotTrees} plot eigen values into a ".pdf" file.}
\usage{
plotTheTree(EnsList, index, ...)
}
\arguments{
\item{index}{an integer. index of which ensemble matrix you want to plot}

\item{ensList}{a list in which elements are ensemble matrices.}
}
\value{
a tree plot
}
\description{
generate tree plots for selected ensemble matrix
\code{plotTrees} plot eigen values into a ".pdf" file.
}

