% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotTrees.R
\name{plotTrees}
\alias{plotTrees}
\title{generate tree plots for each ensemble matrix
\code{plotTrees} plot eigen values into a ".pdf" file.}
\usage{
plotTrees(EnsList, name = "forest", ...)
}
\arguments{
\item{name}{a character vector of length 1. name of the exported .pdf file.}

\item{ensList}{a list in which elements are ensemble matrices.}
}
\value{
a pdf file in the working directory containing all tree plots.
}
\description{
generate tree plots for each ensemble matrix
\code{plotTrees} plot eigen values into a ".pdf" file.
}
\examples{
Sim <- as.simMat(myData)
temperatures <- temperatureSample(start = 0.01, end = 20, n = 20, method = 'random')
\dontrun{
# for illustration only. skip CRAN check because it ran forever.
Ens_list <- getEnsList(Sim, temperatures, MaxIt = 1000, m = 5)
}
\dontshow{
# for CRAN check only
Ens_list <- getEnsList(Sim, temperatures, MaxIt = 5, m = 5)
}
plotTrees(Ens_list, name = "manyTrees")
}

