% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/practiceData.R
\docType{data}
\name{practiceData}
\alias{practiceData}
\title{Grooming network data}
\format{A data frame with 1595 rows and 2 variables:
\describe{
  \item{Initiator}{Grooming Initiator ID}
  \item{Recipient}{Grooming Recipient ID}
  ...
}}
\usage{
practiceData
}
\description{
A dataset containing grooming edgelist among monkeys.
}
\keyword{datasets}

