% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature_select.R
\name{temperatureSample}
\alias{temperatureSample}
\title{generate temperatures
\code{temperatureSample} ADD DESCRIPTION FOR TEMPERATURE~~~~}
\usage{
temperatureSample(start = 0.01, end = 20, n = 20, method = "random")
}
\arguments{
\item{start}{a numeric vector of length 1, indicating the lowest temperature}

\item{end}{a numeric vector of length 1, indicating the highest temperature}

\item{n}{an integer between 10 to 30, indicating the number of temperatures (more explanations on what temperatures are).}

\item{method}{a character vector indicating the method used in selecting temperatures.
It should take either 'random' or 'fixedInterval', case-sensitive.}
}
\value{
a numeric vector of length n representing temperatures sampled.
}
\description{
generate temperatures
\code{temperatureSample} ADD DESCRIPTION FOR TEMPERATURE~~~~
}
\examples{
Sim <- as.simMat(myData)  # as.simMat checked.
temperatures <- temperatureSample(start = 0.01, end = 20, n = 20, method = 'random')
}

