% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEigenvalues.R
\name{getEigenvalueList}
\alias{getEigenvalueList}
\title{generate eigenvalues for all ensemble matrices
\code{getEigenvalueList} get eigenvalues from ensemble matrices}
\usage{
getEigenvalueList(EnsList)
}
\arguments{
\item{EnsList}{a list of ensemble matrices}
}
\value{
a list of \code{eigenvalues} for each of the ensemble matrix in the ensemble matrices list.
}
\description{
generate eigenvalues for all ensemble matrices
\code{getEigenvalueList} get eigenvalues from ensemble matrices
}

