% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEns.R
\name{getEnsList}
\alias{getEnsList}
\title{generate temperatures
\code{getEnsList} get ensemble matrices from given similarity matrix at all temperatures}
\usage{
getEnsList(simMat, temperatures, MaxIt = 1000, m = 5)
}
\arguments{
\item{simMat}{a similarity matrix}

\item{temperatures}{temperatures selected}

\item{MaxIt}{number of iterations for regulated random walks}

\item{m}{maxiumnum number of time a node can be visited during random walks}
}
\value{
a numeric vector of length n representing temperatures sampled.
}
\description{
generate temperatures
\code{getEnsList} get ensemble matrices from given similarity matrix at all temperatures
}
\examples{
Sim <- as.simMat(myData)  # as.simMat checked.
temperatures <- temperatureSample(start = 0.01, end = 20, n = 20, method = 'random')
\dontrun{
# Note: It takes a while to run the getEnsList example.
Ens_list <- getEnsList(Sim, temperatures, MaxIt = 1000, m = 5)
}
\dontshow{
# for CRAN check only
Ens_list <- getEnsList(Sim, temperatures, MaxIt = 5, m = 5)
}
}

