% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrees.R
\name{plotCLUSTERS}
\alias{plotCLUSTERS}
\title{generate tree plots for each ensemble matrix
\code{plotTrees} plot eigen values into a ".pdf" file.}
\usage{
plotCLUSTERS(EnsList, mfrow, mar = c(1, 1, 1, 1), line = -1.5, cex = 0.5,
  ...)
}
\arguments{
\item{EnsList}{a list in which elements are ensemble matrices.}

\item{mfrow}{A vector of the form c(nr, nc) passed to \code{\link{par}}.}

\item{mar}{plotting parameters with useful defaults (\code{\link{par}})}

\item{line}{plotting parameters with useful defaults (\code{\link{par}})}

\item{cex}{plotting parameters with useful defaults (\code{\link{par}})}

\item{...}{further plotting parameters}
}
\value{
a pdf file in the working directory containing all tree plots.
}
\description{
generate tree plots for each ensemble matrix
\code{plotTrees} plot eigen values into a ".pdf" file.
}
\examples{
Sim <- as.simMat(myData)
temperatures <- temperatureSample(start = 0.01, end = 20, n = 20, method = 'random')
\dontrun{
# for illustration only. skip CRAN check because it ran forever.
Ens_list <- getEnsList(Sim, temperatures, MaxIt = 1000, m = 5)
}
\dontshow{
# for CRAN check only
Ens_list <- getEnsList(Sim, temperatures, MaxIt = 5, m = 5)
}
plotCLUSTERS(EnsList = Ens_list, mfrow = c(2, 10), mar = c(1, 1, 1, 1))

}

