% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotEigenvalues.R
\name{plotMultiEigenvalues}
\alias{plotMultiEigenvalues}
\title{plot eigenvalues
\code{plotMultiEigenvalues} plot eigen values into a ".pdf" file.}
\usage{
plotMultiEigenvalues(Ens_list, mfrow, mar = c(2, 2, 2, 2), line = -1.5,
  cex = 0.5, ...)
}
\arguments{
\item{mfrow}{A vector of the form c(nr, nc) passed to par{graphics}.}

\item{mar}{margin parameter}

\item{line}{}

\item{cex}{}

\item{...,}{[link: graphical parameters] such as col, see [link: par].}

\item{eigenvalue_list}{a list in which elements are numeric vectors representing eigenvalues.}
}
\value{
a pdf file in the working directory containing all eigenvalue plots
}
\description{
plot eigenvalues
\code{plotMultiEigenvalues} plot eigen values into a ".pdf" file.
}
\examples{
Sim <- as.simMat(myData)
temperatures <- temperatureSample(start = 0.01, end = 20, n = 20, method = 'random')
\dontrun{
# for illustration only. skip CRAN check because it ran forever.
Ens_list <- getEnsList(Sim, temperatures, MaxIt = 1000, m = 5)
}
\dontshow{
# for CRAN check only
Ens_list <- getEnsList(Sim, temperatures, MaxIt = 5, m = 5)
}

plotMultiEigenvalues(Ens_list, mfrow = c(10, 2), mar = c(1, 1, 1, 1))
}

