% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEigenvalues.R
\name{plotMultiEigenvalues}
\alias{plotMultiEigenvalues}
\title{plot eigenvalues
\code{plotMultiEigenvalues} plot eigen values into a ".pdf" file.}
\usage{
plotMultiEigenvalues(Ens_list, mfrow, mar = c(2, 2, 2, 2), line = -1.5,
  cex = 0.5, ...)
}
\arguments{
\item{Ens_list}{a list in which elements are numeric vectors representing eigenvalues.}

\item{mfrow}{A vector of the form \code{c(nr, nc)} passed to \code{\link{par}}.}

\item{mar}{plotting parameters with useful defaults (\code{\link{par}})}

\item{line}{plotting parameters with useful defaults (\code{\link{par}})}

\item{cex}{plotting parameters with useful defaults (\code{\link{par}})}

\item{...}{further plotting parameters}
}
\value{
a \code{pdf} file in the working directory containing all eigenvalue plots
}
\description{
plot eigenvalues
\code{plotMultiEigenvalues} plot eigen values into a ".pdf" file.
}
\examples{
Sim <- as.simMat(myData)
temperatures <- temperatureSample(start = 0.01, end = 20, n = 20, method = 'random')
\dontrun{
# for illustration only. skip CRAN check because it ran forever.
Ens_list <- getEnsList(Sim, temperatures, MaxIt = 1000, m = 5)
}
\dontshow{
# for CRAN check only
Ens_list <- getEnsList(Sim, temperatures, MaxIt = 5, m = 5)
}

plotMultiEigenvalues(Ens_list = Ens_list, mfrow = c(10, 2), mar = c(1, 1, 1, 1))

}

