% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrees.R
\name{plotTheCluster}
\alias{plotTheCluster}
\title{generate tree plots for selected ensemble matrix
\code{plotTrees} plot eigen values into a ".pdf" file.}
\usage{
plotTheCluster(EnsList, index, ...)
}
\arguments{
\item{EnsList}{a list in which elements are ensemble matrices.}

\item{index}{an integer. index of which ensemble matrix you want to plot}
}
\value{
a tree plot
}
\description{
generate tree plots for selected ensemble matrix
\code{plotTrees} plot eigen values into a ".pdf" file.
}

